+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [NEXT]
 
+### Fixed
+
+- Fixed cannot be parsed exception with nested arrays in credential subject data update
+
 ## [3.2.3] - 2026-06-10
 
 ### Fixed