+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed cannot be parsed exception with nested arrays in credential subject data update
 
+## [3.2.4] - 2026-06-12
+
+### Fixed
+
+- update generic-java-lib to 1.6.4 for didresolver security updates.
+
 ## [3.2.3] - 2026-06-10
 
 ### Fixed